DAOS-8491 vos: Fix non-EC SINGLE gsizes
See

  https://daosio.atlassian.net/browse/DAOS-8491?focusedCommentId=85686

for a description of the problem. In short, if vos_update_begin/end are
used to update a SINGLE value with an iod whose iod_recxs points to a
daos_recx_t object, then the ir_gsize of this value will become the
memory address of the daos_recx_t object. When this SINGLE value is
fetched later, the output iod_size will contain an incorrect size---the
aforementioned memory address! With dss_enum_unpack, the problem affects
at least rdb use cases.

To make the VOS API safer in this regard, this patch changes akey_update
to treat iod_recxs as "gsize" only if the object is an EC object. Also:

  - Add an iod_size verification to vts_io.c. This verification fails
    without the rest of this patch.

  - Add comments mentioning the existing EC iod_recxs hack to VOS update
    APIs.

Signed-off-by: Li Wei <wei.g.li@intel.com>