CORCI-1081 ci: Move Unit Testing from EL7 to EL8
In order to use Dockerfile.centos.8 to build DAOS to be used in Unit
Testing, on our older docker hosts, the DRY mechanism of using FROM
with an argument had to be collapsed.  The unfortunate result is the
need for duplicate EL8 (Rocky, Alma, and Fedora even) Dockerfiles.

The CentOS 8 Dockerfile is being held back to 8.4.2105 until assessment
of how the release of 8.5.2111 will impact the 2.0 release is completed.

Also includes work for:
DAOS-9112 Remove group repos from EL8 test images to tear down the group
repos on EL8 and replace them with their constituents.  This is needed
in particular for the EL8.4 images that are on the NLT hosts.

Skip-PR-comments: true

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>