DAOS-8040 control: show vmd backing addresses in storage scan
Consolidate the scan result filter logic, said logic restricts results
to only devices in the request list and will return error if controller
details for any member of the list is missing.

In the case where VMD addresses have been specified in the request
device list, check that details of the backing device(s) for the VMD
are contained in the scan results and return them in place of the
requested logical VMD address. Return error if no backing devices data
exists for a given VMD address.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>