DAOS-8190 dfuse: Add checking for new fuse flags.
Check for FUSE_CAP_CACHE_SYMLINKS, FUSE_CAP_NO_OPENDIR_SUPPORT and
FUSE_CAP_EXPLICIT_INVAL_DATA in the want flags so we can log
if the kernel supports them.

Downgrade the message for unknown flags from a error to a warning.

There is no code change here, dfuse simply logs the features that
libfuse is offering, the issue is that a newer kernel and libfuse
had set some feature bits that dfuse was not aware of, this change
simply logs the feature bits by name on startup.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>