dfuse: Use seperate hash table buckets for directories.
Use one bit of the inode number to identify something as
a directory, then use this in the hash function to put
directories in dedicated buckets.  This should allow faster
lookups when searching for something that we know is a
directory.

For unlink where we only have the dfs object id then search
both possible inodes which will result in slower unlink times
for directories.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>