DAOS-8559 vos: Implement simplified vos_obj_discard
Rather than trying to fix the ilog, which proves particularly difficult,
we make the simplifying assumption that object discard just removes
values with the expectation that any ilog entries that are left
will have equivalent ilog entries added by reintegration.   If this
is not the case, this function can result in orphaned data.

It is expected that aggregation is disabled during this reintegration
process to avoid potential race conditions that are difficult to
handle.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>