DAOS-8541 metrics: Calculate size for per-pool dir
There are a number of per-pool metrics that are created per-target.
The size for the ephemeral per-pool metrics region is now calculated
based on the reported sizes for each module's per-pool metrics. When
adding metrics to a module, the function pointer .dmm_nr_metrics
should be a function that returns the total number of metrics to be
allocated per-pool by that module.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>