DAOS-8460 vmd: Check for distro for lspci path in detectVMD() and remove device ID (#6609)
Currently VMD devices are queried via lspci, and on OpenSUSE no VMD devices are
returned as the command fails. $lspci is not in the default $PATH for non-root users on
OpenSUSE and SLES. Add check for distro, and if either opensuse, sles, or opensuse-leap,
then use /sbin/lspci.
Also remove VMD PCI device ID (0x201d) from query as the ID is specific to SKX
platforms. Instead only query for "Volume Management Device", as that
remains the same in ICX systems when parsing for VMD addresses.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>