DAOS-8591 test, container: enable default label in daos_test
Use default label for containers in daos_test when a label isn't
passed via the properties.

Also with this change, modify container service code that renames
a container (via set prop) so that in the cs_uuids KVS the new label
akey insert is done before punching the old label akey. When there is
a single entry in cs_uuids, this approach avoids a momentary empty KVS
in the same epoch that would cause vos punch propagation (on the dkey).
And avoids vos returning -DER_NO_PERM due to punch and update.

Change vos_obj_punch() logging to print epoch in hex, aligning it
with other vos logging.

Co-authored-by: Johann Lombardi <johann.lombardi@intel.com>
Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>