DAOS-8542 object: missing bulk free in obj_comp_cb
If no_retry is set for the task, which is usually used by migration
task, then it should not check io_retry, otherwse both io_retry and
no_retry will be set, and retry_cb will not be scheduled, so bulks
will not freed, but its buffer allocated by DAOS will be freed, so
later if the data arrive, it will cause segfaults.

Missing bulk free if retry_cb() failed.

Signed-off-by: Di Wang <di.wang@intel.com>