DAOS-8641 dtx: avoid noisy error log message (#6809)
During rebuild, DTX refresh may hits -DER_STALE when
locate the leader for non-rebuilt EC object. That is
not fatal and will cause related request sponsor to
repeatedly retry until rebuild done or hit some other
failure. Such repeatedly retry may generate huge of
log message as to the server out of space. Since it
is not fatal, convert it from "ERROR" to "TARCE" in
the log.

Signed-off-by: Fan Yong <fan.yong@intel.com>