DAOS-8586: container: snapshot commands output hex epochs (#6803)
In a previous change, the daos container list-snap command was changed
so its human-readable output presented epochs in hex format. However,
other commands such as create-snap continued to output base 10 values.
This was an oversight. And, the container destroy-snap command --epc
and --epcrange arguments did not properly parse hex formatted epochs.

With this change, all snapshot human-readable output is changed so that
epochs are printed in hex format with a 0x prefix. This includes
container query (ci_lsnapshot most recent snapshot). And the snapshot
commands accepting --epc or --epcrange arguments accept either decimal
or hexadecimal formatted values.

In a future change, the daos snapshot commands, when run with the -j
argument to the daos utility, will produce JSON-formatted information.
Currently the tools call a C-based handler function that produces
only human-readable output. When that change is made, daos -j callers
will then be able to take the unformatted value and pass it as the
--epc argument to other snapshot commands.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>