dfuse: Test for and resolve issues in starting dfuse.
Add a fault injection test for dfuse and resolve any issues
to ensure that there are no secondary errors reported and
everything is closed properly.

More work is needed here in various areas, daos_pool_close()
can sometimes work but return an error code so on the second
call it returns another error.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>