DAOS-623 vos: Remove some dead code
Before the incarnation log, we used direct keys to implement integer
and lexical keys because the actual key consisted of the key and the
epoch.   With incarnation log, we only need to store the key so
we use built-in btree functionality for lexical and integer keys.

This just removes some old, confusing dead code associated with
this change

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>