DAOS-8693 dtx: simplify DTX retry logic
Originally, for read style RPC, if hit sync mode non-committed DTX,
we can make the RPC handler (ULT) to re-schedule locally instead of
asking the client to resend the RPC. It is expected that such sync
mode DTX will be committed soon. But it is not guaranteed. If such
sync mode DTX is still not committed after several cycles of local
re-schedule, we will have to make the client to resent the RPC (to
the leader if possible). So this 'optimization' cannot replace the
client retry logic. On the other hand, current test results cannot
prove sich 'optimization' is worth for the relative complex servre
retry logic. Let's remove server local retry to simplify the logic.

Signed-off-by: Fan Yong <fan.yong@intel.com>