DAOS-8445 tests: not set EC cell size on container for racer (#6835)
It is unnecessary to explicitly set container's EC cell size
as 1M for racer test. Because:

If the user specifies the container to be used for racer test,
then we have no way to change an existing container's EC cell
size. Or

If the user does not specify the test container, then the new
created (by racer) container will inherit such property from
the pool which is 1M by default.

Signed-off-by: Fan Yong <fan.yong@intel.com>