DAOS-8694 tests: Updating expected behavior of server_name_junk
In the control/daos_agent_config.py test the server_name_junk varaint,
which uses "! @#$%^&*()_+{}|:<>?-=[];',./" as the server_name value in
the daos_agent config file, the daos_agent was expected to start w/o an
error.  Now the agent fails to start with this invalid value and the
test is being updated to expect this behavior.

Quick-functional: true
Test-tag: test_daos_agent_config_basic

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>