DAOS-7609 dtx: not abort DTX entry when DTX refresh
When non-leader refresh some DTX staus with the leader, it may
get the result -DER_NONEXIST. There are three cases for that:

1) The leader has aborted related DTX but the non-leader missed
   or failed to executed the DTX abort RPC.

2) Such DTX entry has been committed, but the non-leader missed
   or failed to executed the DTX commit RPC. Then it is removed
   by DTX aggregation on the leader.

3) The RPC handler on the leader yield CPU before related local
   DTX entry pinned because of NVMe data copy/movement. Refresh
   such DTX from the leader will not find it.

Currently, we do not have efficient way to distinguish these cases.
So we cannot abort the DTX entry for -DER_NONEXIST case. Otherwise
it may break leader's local modification or commit.

Signed-off-by: Fan Yong <fan.yong@intel.com>