DAOS-8699 container: Use UV class for container btree
The key is a uuid, so it's appropriate to use the UV class instead.
The NV class without direct key can have hash conflicts between
two different UUIDs.   This is similar to what we saw previously
with objects in DAOS-6149.   It's likely that the customer is just
using more containers than we're used to seeing.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>