DAOS-8559 vos: Disable range discard on ilog
Just discard the ilog entries in the range instead.

Not doing so can leave data that is punched but the punches
may have been aggregated and can't be reintegrated.  Also,
supporting inflight I/O with discard is possibly impossible.

The issue with in-flight I/O in such cases is supporting
conditional operations when you have potentially stale
I/Os at earlier epochs.

So, we are punting and just reverting to full discard of the
ilog and assuming no I/O can happen on the such objects
after restart but before discard has completed.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>