DAOS-7649 daos: add serialize and deserialize utility with tests
Adds a daos serialize and daos deserialize utility. The daos
serialize utility writes the daos container to an HDF5 file,
and the daos deserialize tool takes in the HDF5 file and writes
the data back into a new daos container. The HDF5 specific code
is isolated to the src/client/serialize/daos_serialize.c file.

Test-tag: pr cont_clone fs_copy cont_serialize

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>