DAOS-8007 Test: Fix VMD test identification.
There is an issue to detect the proper backing NVMe device for VMD.
Current cut options does not work on some cases so updated the cut
option to avoide that issue and work on each case.

PR will be tested on VMD enabled HW so no other testing needed in
ci.

Signed-off-by: Samir Raval <samir.raval@intel.com>