DAOS-8773 dtx: restart distributed transaction if pool map refreshed
On the client, when prepare CPD RPC for the distributed transaction,
we need to check whether the pool map is freshed after the transaction
started or not. If yes, we need to restart the transaction, because the
layout for the touched objects may have been changed and related time
stamps for the distributed transaction may be lost on related servers.

Signed-off-by: Fan Yong <fan.yong@intel.com>