DAOS-8835 vos: not clean non-exist DXT entry
The DTX entry may has been aborted during current ULT waiting
for some event, such as bulk data transfer. Then when current
ULT try to cleanup the DTX entry for some reason, it needs to
check whether related DTX entry is still in the active table
or not, if not then skip the cleanup.

Signed-off-by: Fan Yong <fan.yong@intel.com>