DAOS-7069 dtx: properly abort DTX entry when DTX refresh
master-commit: 34f072cf20da32f3b9baf492837881542ea45e16

When non-leader refresh some DTX staus with the leader, it may
get the result -DER_NONEXIST. There are two cases for that:

1) The leader has aborted related DTX but the non-leader missed
   or failed to executed the DTX abort RPC.

2) Such DTX entry has been committed, but the non-leader missed
   or failed to executed the DTX commit RPC. Then it is removed
   by DTX aggregation on the leader.

If we are not sure which case it is, then we will have to keep
related of DTX entry instead of aborting it. Otherwise, we may
break leader's local modification or commit.

The patch also contains other fixes to allow cleanup 'prepared'
DTX entry.

Signed-off-by: Fan Yong <fan.yong@intel.com>