pass parent default to create dir that could be set in yaml file
The parent directory that can now be set in the yaml was not
being passed to new_posix_test_dir. This allows someone
to change parent directory without modifying the test files.
The default for parent is set to self.tmp and on the large
hardware clusters this uses a shared posix directory
(i.e. not tmpfs).

Test-tag: datamover
Quick-Functional: true

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>