DAOS-8826 dtx: stop current batched commit if container re-opened
For each container open instance, there will be a 'dbca' structure
registered to the batched commit main ULT. Then the batched commit
main ULT will create child ULT for such container instance if have
something to be batched committed. To be safe the child ULT should
exit if the container is closed and re-opened to avoid the case of
multiple child ULTs to do batched commit for the same container.

Signed-off-by: Fan Yong <fan.yong@intel.com>