DAOS-8877 object: simplify CPD handler for single TX based CPD RPC
Our CPD RPC is designed to support multiple transactions via the
same CPD RPC. On the leader server, for each transaction in the
CPD RPC, the RPC handler will create an ULT to handle it. Then
all these transactions can be processed in parallel. But if the
CPD RPC only contains one transcation, then we can directly handle
it via current RPC handler instead of creating another ULT. That
is optimization for current distributed transaction.

Signed-off-by: Fan Yong <fan.yong@intel.com>