Update setup.md
Updated some of the links in here that pointed to 1.2 branch and github.daos.io vs daos.docs.io.
The RPM installation section is problematic, as the links there should point to the 'Dev' location that has not yet been populated. Let me know how you want to handle that.