DAOS-8782 pool: Remove meta files (#6949)
The "meta" file in each pool's directory store the "target UUID" of the
engine in this pool. The file and the UUID were originally introduced
for engine ranks that could change across system restarts. Since ranks
are now persistent, these files and UUIDs are no longer required.
Moreover, the UUIDs are also stored in the pool RDB, but they do not
include any UUIDs corresponding to engines added to the pool after the
pool is created. Hence, this patch removes these UUIDs and "meta" files
to simplify the persistent state as well as to avoid confusions:

  - Determine node component's IDs in a pool map using ranks instead of
    "target UUIDs".
  - Remove "target UUIDs" in pool RDBs.
  - Remove "target UUIDs" in pool RPCs.
  - Remove "meta" files.

Signed-off-by: Li Wei <wei.g.li@intel.com>