DAOS-8823 dfuse: Use correct permissions bits in readdir. (#6992) (#7050)
Correct a bug in readdir where dfs handles were being
opened read only.  If the handle was not already in the
hash table then the readdir one would be added, leading
to an inability to modify the file afterwards.  To reproduce
it was necessairy to create a file, restart dfuse then
access the file through readdir rather than lookup/stat
at which point subsequent setattr calls would fail.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>