DAOS-8433 dtx: allocate DTX handle dynamically
Current, the size for dtx_leader_handle is more than 300 bytes, that
is some large for being as static variable on ULT stack, and may be
one of the important factors of ULT stack overflow. In the patch, we
dynamically allocate the DTX handle via dtx_leader_begin(), and free
it via dtx_leader_end(). Similar for dtx_begin() and dtx_end().

Signed-off-by: Fan Yong <fan.yong@intel.com>