DAOS-8685 test: configure default nr_xs_helpers: 4
Cherry-pick PR #7040 master commit b7311ec to release/2.0 branch.

Before this change, if a functional test .yaml file did not specify
nr_xs_helpers, the server_utils_params.py code would set it to 16.
This would result in (along with a default of 8 targets), at least
24 threads (8 target pthreads, 16 helpers) along with the usual
system xstream and swim xstream threads. If run on a (for example)
single CPU of a dual-CPU system with 18 physical cores per CPU,
there is concern for oversubscription and possible CPU scheduling
contention - that may be more pronounced in particular OS / kernel
versions.

With this change, the default setting is 4 helpers, one helper thread
responsible for 2 (of the default 8) targets.

Test-tag: pr daily_regression metadata_fillup metadata_addremove cpu_usage

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>