DAOS-8915 build: Build and test on Leap 15.3
Remove build stages duplicted in GHA to make more room in the
Jenkinsfile for other required stages.

Force Leap RPM test stage to use 15.2 so that we have coverage for that
platform as we move main testing to Leap 15.3.

Leap 15.3 seems to be defaulting to kernel.dmesg_restrict=0 so reverse
that so that we can gather dmesg without being root.

Coalesce sysctl settings into one place.  Simplify the logic for finding
ib* interfaces and including them in the sysctl settings.

Update packaging to pull in changes to allow setting distro-setting
pragmas such as "LEAP15-target: leap15.3" and LEAP15-version: 15.3.

Skip-PR-comments: true
Skip-func-test-leap15: false

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>