DAOS-8672 dtx: trace the newest DTX that is aggregated (#7119)
master-release: 440d04127110efdeeb9278866ed7e4aeef69a279

Record the epoch for the newest DXT entry that is aggregated by the
DTX aggregation. Such epoch will be used when check resent RPC and
DTX refresh: if we cannot find the DTX with the give transaction ID
and if the DTX timestamp is newer than such newset aggregated epoch,
then it is impossible that it has been removed by DTX aggregation.

The patch also increases the default value for DTX aggreagtion age
threshold. Then if there is no DRAM pressure on server, the committed
DTX entries can be kept for more long time. That is helpful for resend
logic to know whether related DTX has even been committed. But if there
is DRAM pressure on the server, then DTX aggregation will be triggered
via the count threshold.

Signed-off-by: Fan Yong <fan.yong@intel.com>