DAOS test: Increase expected logfile size for NLT.
NLT enforces a maximum log-file size to catch cases where
we acidentally increase the logging verbosity, however
as we've added more tests the logs have got bigger so we're
now failing the size check randomly.

Increase the size to 10% so that it's within expected
constraints now, but we still pick up major regressions.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>