DAOS-8888 misc: Fix ULT leaks (#7098)
Argobots dictates the following:

  - A ULT that is not unnamed must be freed explicitly via
    ABT_thread_free. (ABT_thread_join does not do this.)
  - An unnamed ULT cannot be freed explicitly via ABT_thread_free.
  - A ULT cannot free itself.
  - ABT_thread_free implicitly joins the ULT before freeing it.
  - An unnamed ULT cannot be joined via ABT_thread_join.

In several modules, ULTs are only joined but not freed, leading to ULT
leaks. This patch replaces these ABT_thread_join calls with
ABT_thread_free call, which both join as well as free. Also, in some
error paths among those cases, the ULTs may not terminate, leading to
hangs. This patch adopts a check-if-req-is-null-in-the-ULT approach from other
error paths, whose correctness subtly relies on the absence of yielding
between the ULT creation and the ABT_thread_free call---something that
should be replaced with a more robust solution in the future.

sched_req_put does not free the ULT too. This patch adds an "owned" flag
to sched_req so that sched_req_put can call ABT_thread_free on ULTs that
are req-owned. Some dtx ULTs' sched_req_put calls have to be adjusted a
bit to avoid those ULTs from freeing themselves.

Signed-off-by: Li Wei <wei.g.li@intel.com>