DAOS-7878 vos: Address punch propagation performance
* Rather than potentially iterating all keys twice during punch
propagation, iterate once and track whether we've seen
entries or uncommitted data.
* Use existing, unused space in the object and dkey durable format
to save offset of last key known to exist so we can skip scan.  This can
significantly speed up punch propagation in the common case by limiting
full scan to the occasional instance where the saved key is punched.
for punch propagation if the key hasn't been punched.
* Enable punch propagation to object level and associated mvcc tests
* Always start scan using prior key as anchor

In vos_perf tests, I got the following results for punch of 100K dkeys
With this patch:                                0.7 seconds
with master branch:                             0.6 seconds
master branch with propagation to object level: 118 seconds

Seems like a good tradeoff.  It should improve delete performance of
directories containing many objects as in the ticket.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>