DAOS-8987 vos: Combine emptiness and uncommitted checks into one
Rather than potentially iterating all keys twice during punch
propagation, iterate once and track whether we've seen
entries or uncommitted data.

Also, use existing, unused space in object and dkey durable format
for saving information for future use that will enable skipping
full propagation check most of the time.  Reset this data if set
so we are forward compatible when this is implemented. Bump the
version so we use this forward compatible version

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>