DAOS-8976 dfuse: Fix return codes if ioctl calls fail.
Fix a couple of bugs in dfuse where if a call to dfs global
to local failed then the wrong error could would be returned,
a system error code would be assumed to be a daos one and
converted again.

Add a extra check for one place where the return value
was not being checked.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>