DAOS-9017 dfuse: Add checking for fetching handles for non-files. (#7203)
If a non-file is opened by the IL then we currently get an error
in dfuse from the ioctl, and a error-path is taken in the IL
that will cause it to disconnect from the container.

After this patch we take the following behaviour:
When a directory is opened the container/pool are opened and
dfs is mounted, however when it comes to opening the dfs object
the ioctl code in dfuse will return with E_ISDIR rather than
EINVAL and the IL itself will spot this and not attempt to
release resources - so it will keep dfs mounted for any future
access.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>