DAOS-9097 test: Allow the build to continue on valgrind error.
Change the behaviour of NLT not to exit non-zero if there are
valgrind issues reported.  These errors are flagged by Jenkins
when the build completes stopping the build is not required,
rater have NLT add them as a test failure, then it's up to the
build configuration if jobs stop or not.

Also detect the case where exceptions mean that an attempt
is made to stop the server whilst dfuse instances are running,
if this happens then stop the dfuse instance before stopping
the server, and again register this as a test failure.

Fix a logparse bug where a recent change meant that client
logs over 100Mb in size could not be parsed,

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>