DAOS-9115: socket dirs and systemd RuntimeDirectory should match
The daos_server.yml socket_dir and daos_agent.yml runtime_dir settings
for the socket directories should be kept at their defaults, especially
if the daemons are under systemd control. The systemd RuntimeDirectory
controls the location of these socket directories, and systemd will
create/remove them with the right permissions when the services are
started/stopped.

The example YML files and documentation have been updated with a warning
that these must match the systemd RuntimeDirectory (when daemons are under
systemd control) and/or those lines have been removed from the examples.

Signed-off-by: Michael Hennecke <michael.hennecke@intel.com>