DAOS-9149 duns: simple xattr format and all cont types support
At the beginning of DUNS design, the DUNS xattr format was intended
to contain extra informations like container-type, object-type,
chunk-size, ...
Most of them are no longer stored this way and thus we can simplify
the DUNS xattr format on Lustre file-system to avoid the need of a
user-helper to decode it by default.
The idea is to only store a relative path made of "<PUUID>/<CUUID>"
in the foreign LOV/LMV free-format lfm_value field, and
the container-type in lfm_flag field.
Thus to construct the full foreign_symlink destination, the
Lustre/llite module will have to simply concatenate the
foreign_symlink_prefix with the current foreign file/dir lim_value
without any other operation.

Also, the support of all container types in DUNS/Lustre integration
was missing, so this patch also implements it.

Change-Id: I48c192581b10ffa7fcd65962c73c5a84d557e7c3
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>