DAOS-7821 gha: Perform fewer builds for pull requests.
Cut down the number of distros that we build for PRs, remove ci.yml
completely.

Now we're building on master again we don't need to cover so many
distros with PRs, and github has been queueing a number of our
builds recently.

Set the target branch in the hash key for docker images.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>