DAOS-9103 test: Change daos_test to use --nsvc=3
After the change to randomize pool service ranks landed,
the degraded mode tests started randomly failing. The
problem is that the test always tries to kill ranks 7,6,5
without regard for where the pool service is running.
The test does avoid killing the pool service rank, but then
the test fails because it expects all three ranks to be in
the excluded state at the end.

The primary fix for this problem is to create pools with
a resilient pool service when ranks will be killed so that
it can survive the failure of a rank. The secondary fix is
to modify the daos_kill_server() helper to allow killing of
svc ranks unless there is only a single svc rank.

Test-tag: pr test_daos_degraded_mode test_daos_oid_allocator

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>