DAOS-9181 EC: only select last parity shard as EC agg leader
To avoid both parity shards treat the other is leader when they with
different pm version.
For example, 4P2G1 EC obj, with shards 0~5,on server of shard 4 with
pool map ver 10 all are alive and it check shard 5 should be EC agg
leader so shard 4 will do nothing; at that time, on server of shard 5
with pool map ver 11, it checks itself failed or in-rebuilding, then it
think shard 4 should be the leader so shard 5 will do nothing.
Both servers will not prevent the EC agg epoch move ahead but actually
did not do EC agg work.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>