DAOS-7821 gha: Cleanup new github actions jobs.
Rename landings jobs so they display in the web interface properly.
Run landings jobs in PRs that modify them
Run the NLT test for all distros.
Increase the max-parallel setting to speed up master jobs

Previously we saw long waits for GHA jobs to be scheduled, to
mitigate this we run fewer jobs on PRs but also rate limited
master jobs.  This worked so well that often now the only jobs
running are the rate-limited master jobs so turn back up the
speed that these can run.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>