DAOS-9189 Test: Fix the common code JSON to not to use the loop.
On CentOS 8.4 which has json package 0.13.1-0.4.On latest changes
json_object_object_get_ex() fails if the value is empty. And in device-health
pools is key with no value which is getting Assertion on latest JSON.

So removed the for loop and instead use the value of device and read
health info.

Signed-off-by: Samir Raval <samir.raval@intel.com>