Fix -DER_NOMEMs from raft_add_node
When re-adding an RDB replica, we may observe an -DER_NOMEM error from
raft_add_node, because, for example:

  Index  Change
  --------------------
      1  Add rank 1-7
     12  Add rank 0
     13  Remove rank 7
     32  Add rank 7
     33  Remove rank 6

When rank 7 processes entry 13, it doesn't remove the raft_node_t
object. Therefore, when entry 32 re-adds rank 7, the existing
raft_node_t leads to an error. We attempt to fix this by removing the
raft_node_t object when removing rank 7.

Signed-off-by: Li Wei <wei.g.li@intel.com>