DAOS-8346 test: Use context manager for daos server in NLT.
Avoid a bug where large log files could be processed twice, leading
to increased runtime.  What was happening was the object was handling
the file in stop(), and again in __del__() if it still existed, and
because the bzip2 process was run in the background it might have been,
plus because of the way python handles objects there's very little
control over when __del__ is called.

Remove the logfile from the object once it's processed as a first-off
which will avoid the bug, but use DaosServer under a context manager
which should destroy the object in a more timely manner as well.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>