DAOS-9350 log: Fix issues with DF_KEY and daos_key2str
First of all, daos_key2str returns a NULL terminated string
so there is no need to limit the characters we print.

Secondly, the %*s in daos_key2str should be %.*s.  The former
is valid printf but simply is a format specifier whereas the
latter actually limits the characters printed.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>