DAOS-9381 sched: don't delay IO requests by mistake
The IO requests processing design is to ensure that all requests
for different pools are processed in FIFO order, and the space
pressure from one pool shouldn't interfere the requests processing
for other pools. However, there is a defect in policy_fifo_process()
which makes the latter requests (in FIFO queue) could be impeded by
the former request (for a different pool with space pressure).

This patch changed policy_fifo_process() to avoid the defect.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>