DAOS-8482 bio: set hotplug filter with pci bus id range (#7653)
Hotplug events are to be filtered so that an engine only responds to
events on PCI addresses with specific bus-ID values. This is to avoid
the situation where multiple per-engine SPDK hotplug pollers mistakenly
remove/plug the same device.

The hotplug filter configuration can be a hexadecimal bus ID range
e.g. 0-FF and will be automatically calculated during storage format based on
the NUMA node a particular engine is bound to unless an explicit value
set using the bdev_busid_range storage parameter of the server config file.
The range will be set to allow any (0-FF) when VMD is enabled as the SPDK
VMD driver will implicitly provide per-engine device filtering which will affect
hotplug event delivery.

NUMA node of an engine is either explicitly assigned using the
pinned_numa_node parameter of the server config file or derived from the
affinity of the device identified by the fabric_interface server config
parameter.

The range will be communicated to the engine by being written into the
SPDK JSON file that defines the bdev storage to be used by an engine.

The engine's bio module will then decode the range from JSON and
set the hotplug filter function (using the range as decider logic) to be
used to filter events detected by the SPDK hotplug poller.

Related changes:
  * improve error reporting during storage format
  * move pinned_numa_node from fabric to engine config
  * improve mocking of engine config to be used in unit tests
  * improve engine config validation
  * set affinity of engine based on fabric if not set explicitly

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>