DAOS-8682 doc: Added usable space notice
Added the following notice about usable space in
"Creating a Pool" section:

Note that it is difficult to determine the usable space by the user and
currently we cannot provide the precise value. The usable space depends not only
on pool size, but also on number of targets, target size, object class,
redundancy factor, etc.

Skip-build: true
Skip-test: true
Signed-off-by: Makito Kano <makito.kano@intel.com>