DAOS-9635 object: multiple shards enumeration fix
The patch fix the enumeration from multiple data shards,
for example degrade enumeration, or parity rotation,

1. Each shard will prepare KDS/SGL/RECX buffer invidually,
and attached to the anchor.

2. After each shard get the enumeration buffer, it
needs to sort these dkey/recxs from all shards altogether
by its order inside the VOS, so to avoid duplicate/missing
keys for enumeration.

3. Then these merged/sorted keys/recxs are dumped to
the user kds/sgl/recx, and the leftover will be left in
the merged list, which will be retrieved in the following
enumeration.

Signed-off-by: Di Wang <di.wang@intel.com>